+++ autosubmit_api/database/tables.py
@@ -1,51 +1,59 @@
-from sqlalchemy import Table, Column, MetaData, Integer, String, Text
+from sqlalchemy import MetaData, Integer, String, Text, Table
+from sqlalchemy.orm import DeclarativeBase, mapped_column, Mapped
+
 
 metadata_obj = MetaData()
 
 
+class BaseTable(DeclarativeBase):
+    metadata = metadata_obj
+
+
+class ExperimentTable(BaseTable):
+    __tablename__ = "experiment"
+
+    id: Mapped[int] = mapped_column(Integer, nullable=False, primary_key=True)
+    name: Mapped[str] = mapped_column(String, nullable=False)
+    description: Mapped[str] = mapped_column(String, nullable=False)
+    autosubmit_version: Mapped[str] = mapped_column(String)
+
+
+class DetailsTable(BaseTable):
+    __tablename__ = "details"
+
+    exp_id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    user: Mapped[str] = mapped_column(Text, nullable=False)
+    created: Mapped[str] = mapped_column(Text, nullable=False)
+    model: Mapped[str] = mapped_column(Text, nullable=False)
+    branch: Mapped[str] = mapped_column(Text, nullable=False)
+    hpc: Mapped[str] = mapped_column(Text, nullable=False)
+
+
+class ExperimentStatusTable(BaseTable):
+    __tablename__ = "experiment_status"
+
+    exp_id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    name: Mapped[str] = mapped_column(Text, nullable=False)
+    status: Mapped[str] = mapped_column(Text, nullable=False)
+    seconds_diff: Mapped[int] = mapped_column(Integer, nullable=False)
+    modified: Mapped[str] = mapped_column(Text, nullable=False)
+
+
+class GraphDataTable(BaseTable):
+    __tablename__ = "experiment_graph_draw"
+
+    id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    job_name: Mapped[str] = mapped_column(Text, nullable=False)
+    x: Mapped[int] = mapped_column(Integer, nullable=False)
+    y: Mapped[int] = mapped_column(Integer, nullable=False)
+
+
 # MAIN_DB TABLES
-
-experiment_table = Table(
-    "experiment",
-    metadata_obj,
-    Column("id", Integer, nullable=False, primary_key=True),
-    Column("name", String, nullable=False),
-    Column("description", String, nullable=False),
-    Column("autosubmit_version", String),
-)
-
-details_table = Table(
-    "details",
-    metadata_obj,
-    Column("exp_id", Integer, primary_key=True),
-    Column("user", Text, nullable=False),
-    Column("created", Text, nullable=False),
-    Column("model", Text, nullable=False),
-    Column("branch", Text, nullable=False),
-    Column("hpc", Text, nullable=False),
-)
-
+experiment_table: Table = ExperimentTable.__table__
+details_table: Table = DetailsTable.__table__
 
 # AS_TIMES TABLES
-
-experiment_status_table = Table(
-    "experiment_status",
-    metadata_obj,
-    Column("exp_id", Integer, primary_key=True),
-    Column("name", Text, nullable=False),
-    Column("status", Text, nullable=False),
-    Column("seconds_diff", Integer, nullable=False),
-    Column("modified", Text, nullable=False),
-)
-
+experiment_status_table: Table = ExperimentStatusTable.__table__
 
 # Graph Data TABLES
-
-graph_data_table = Table(
-    "experiment_graph_draw",
-    metadata_obj,
-    Column("id", Integer, primary_key=True),
-    Column("job_name", Text, nullable=False),
-    Column("x", Integer, nullable=False),
-    Column("y", Integer, nullable=False),
-)
+graph_data_table: Table = GraphDataTable.__table__