+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Pre-release v4.0.0b6 - Release date: TBD
 
+* Added filter by autosubmit version in `/v4/experiments`
+* Fixed some connection file handlers
+
 ### Pre-release v4.0.0b5 - Release date: 2024-03-18
 
 * Fixed the graph background task to correctly select the active experiments